refactor: general components folder
