+++ src/app/_components/card/mahasiswa-card.tsx
