+++ src/app/_components/layout/navbar.tsx
@@ -16,7 +16,7 @@ import {type Session} from "next-auth";
 // Assets Import
 import LogoAnmategra from "/public/images/logo/anmategra-logo-full.png"
 
-const MahasiswaSidebar = ({ session }: { session: Session | null }) => {
+const Navbar = ({ session }: { session: Session | null }) => {
   const [searchQuery, setSearchQuery] = useState('');
   const router = useRouter();
 
@@ -85,4 +85,4 @@ const MahasiswaSidebar = ({ session }: { session: Session | null }) => {
   );
 };
 
-export default MahasiswaSidebar;
+export default Navbar;
