+++ src/app/mahasiswa/layout.tsx
@@ -1,5 +1,5 @@
 import React, { type ReactNode } from 'react'
-import MahasiswaSidebar from '../_components/MahasiswaSidebar'
+import Navbar from '../_components/layout/navbar'
 import {getServerAuthSession} from "~/server/auth";
 
 const HalamanMahasiswaLayout = async ({ children }: { children: ReactNode }) => {
@@ -7,7 +7,7 @@ const HalamanMahasiswaLayout = async ({ children }: { children: ReactNode }) => 
   return (
     <div className='w-full flex flex-col items-center'>
       <div className='w-full max-w-7xl'>
-        <MahasiswaSidebar session={session} />
+        <Navbar session={session} />
         {children}
       </div>
     </div>
